implement TeammatesAndRequests component
